Add `with_lock` block to ScoreRound service
When adding `with_lock` on a block around Board, it looks like the ScoreRound
spec already has a previous reference to `player1`, and ActiveRecord fails to
recognise this previous reference - so updating the Player within the locked
block fails to update the reference in the spec.

Therefore all `let(:arg) { obj.create! }` should instead be references through
`obj.children.where()`, and the objects are actually created in a `before`
block instead. This can ensure all references are not cached.

This is gross, but it is the cleanest approach we've thought of so far to
both maintain the consistency of the tests, and without scattering handfuls
of `.reload()` calls throughout the specs.